complete dashborad and some other changes

+++ controllers/categoryController.js
@@ -1,5 +1,6 @@
 import categoryModel from "../models/categoryModel.js";
 import transactionTypeModel from "../models/transactionTypeModel.js";
+import transactionModel from "../models/transactionModel.js";
 
 async function addCategory(req, res) {
     try {
@@ -51,6 +52,13 @@ async function addCategory(req, res) {
 async function deleteCategory(req, res) {
     try {
         const categoryId = req.params.id;
+        const hasTransaction = await transactionModel.exists({categoryId: categoryId});
+
+        if(hasTransaction){
+            req.flash("error","Cannot delete category because it has transactions.");
+            return res.redirect("/categories");
+        }
+
         const deleted = await categoryModel.findByIdAndDelete(categoryId);
 
         if (!deleted) {
@@ -70,40 +78,65 @@ async function deleteCategory(req, res) {
 async function editCategory(req, res) {
     try {
         const categoryId = req.params.id;
+        const userId = req.session.user._id;
         const { name, transactionType } = req.body;
 
         if (!name || !transactionType) {
-            req.flash("error", "Name and transaction Type required");
+            req.flash("error", "Name and Transaction Type required");
             return res.redirect("/categories");
-
         }
+
         const type = await transactionTypeModel.findOne({ name: transactionType });
+        if (!type) {
+            req.flash("error", "Invalid Transaction Type");
+            return res.redirect("/categories");
+        }
+
+        const category = await categoryModel
+            .findOne({ _id: categoryId, userId })
+            .populate("transactionType");
+
+        if (!category) {
+            req.flash("error", "Category not found");
+            return res.redirect("/categories");
+        }
+
+        const isNameChanged =
+            category.name.toLowerCase() !== name.trim().toLowerCase();
+        const isTypeChanged =
+            category.transactionType._id.toString() !== type._id.toString();
+
+        if (!isNameChanged && !isTypeChanged) {
+            return res.redirect("/categories");
+        }
 
         const existing = await categoryModel.findOne({
             _id: { $ne: categoryId },
+            userId,
             name: { $regex: `^${name.trim()}$`, $options: "i" },
             transactionType: type._id
         });
 
         if (existing) {
-            req.flash("error", "Category exist");
+            req.flash("error", "Category already exists");
             return res.redirect("/categories");
         }
-        const updated = await categoryModel.findByIdAndUpdate(categoryId, {
-            name: name.trim(),
-            transactionType: type._id
-        }, { new: true }).populate("transactionType");
 
-        req.flash("success", "Catagory updated!");
-        return res.redirect("/categories")
+        category.name = name.trim();
+        category.transactionType = type._id;
+        await category.save();
+
+        req.flash("success", "Category updated!");
+        return res.redirect("/categories");
+
     } catch (error) {
         console.log(`Updating Category Error: ${error.message}`);
-        req.flash("error", "Something went wrong whil`e updating category");
-        return res.redirect("/categories")
-
-
+        req.flash("error", "Something went wrong while updating category");
+        return res.redirect("/categories");
     }
 }
+
+
 
 export default {
     addCategory,
